# docs/openapi.yaml
openapi: 3.0.3
info:
  title: McAforo API
  description: |
    Public + authenticated API for McAforo.
    This spec is the contract consumed by the web app and the upcoming
    Android / iOS mobile apps. All `/api/v1/*` endpoints use a standardised
    error envelope; see `components.schemas.ApiError`.
  version: "1.0.0"
servers:
  - url: https://mcaforo.com
    description: Production
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password (returns bearer + refresh)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1 }
                deviceId: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
        '422':
          description: Validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and get a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
                deviceId:     { type: string, maxLength: 128 }
      responses:
        '200':
          description: Rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Invalid, expired, revoked, or reused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, UNPROCESSABLE, RATE_LIMITED, UPGRADE_REQUIRED, INTERNAL]
        message: { type: string }
        details: {}
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ApiError' }
    SessionUser:
      type: object
      required: [id, email, firstName, lastName, roles]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        roles:
          type: array
          items: { type: string }
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn, tokenType]
      properties:
        accessToken:  { type: string, description: "Short-lived JWT (~15 min)" }
        refreshToken: { type: string, description: "Opaque rotating token (~30 days)" }
        expiresIn:    { type: integer, description: "Access token TTL in seconds" }
        tokenType:    { type: string, example: Bearer }
        user:         { $ref: '#/components/schemas/SessionUser' }
...